unoplat-code-confluence-openmetadata/pyproject.toml
[project]
name = "unoplat-code-confluence-openmetadata"
version = "0.1.0"
description = "OpenMetadata custom ingestion source for Code Confluence repository snapshots"
readme = "README.md"
authors = [
    { name = "JayGhiya", email = "ghiya6548@gmail.com" }
]
requires-python = ">=3.10,<3.13"
dependencies = [
    "cachetools==5.5.2",
    "httpx==0.28.1",
    "openmetadata-ingestion==1.12.6.2",
    "pydantic==2.11.7",
]

[dependency-groups]
test = [
    "pytest==8.3.4",
    "pytest-asyncio>=0.25.0,<0.26",
    "respx==0.22.0",
]
dev = [
    "basedpyright>=1.32.1",
    "ruff>=0.1.0",
]

[build-system]
requires = ["uv_build>=0.11.6,<0.12.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.basedpyright]
typeCheckingMode = "strict"
include = ["src"]
exclude = ["**/__pycache__", ".venv", "dist", "build"]
ignore = ["**/__pycache__"]
useLibraryCodeForTypes = true
reportMissingImports = "error"
reportMissingTypeStubs = false
reportUnusedImport = "warning"
reportUnusedVariable = "warning"
